+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 yii2-widget-cropbox
 ===================
 
-v? [?]
-------
+v5.0.0 [2017-05-07]
+-------------------
 
 - Fixed headers style of markdown. (bryant1410)
+- Removed `bupy7\cropbox\Cropbox` and `bupy7\cropbox\MouseWheelAsset` class.
+- Added `bupy7\cropbox\CropboxWidget` class.
+- Removed `assets` directory.
+- Replaced `bower-asset/jq-cropbox` to `bower-asset/js-cropbox` dependency.
+- Added jQuery wrapper for `js-cropbox` extension.
+- Rewrite `README.md`.
 
 v4.1.2 [2016-09-28]
 -------------------